+++ APIWithControllersAsPlugins/Plugins/OrdersControllerPlugin/Models/OrderTableEntity.cs
@@ -8,7 +8,7 @@ namespace OrdersControllerPlugin.Models
     public class OrderTableEntity : TableEntity
     {
         public int Id { get; set; }
-        public decimal Price { get; set; }
-        public decimal PriceIncVAT { get; set; }
+        public double Price { get; set; }
+        public double PriceIncVAT { get; set; }
     }
 }
